# specs/5-k8s-deployment-oke/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Chatbot API
  description: API for the Todo Chatbot application with advanced features
  version: 1.0.0
servers:
  - url: https://api.todo-chatbot.com/v1
    description: Production server
  - url: http://localhost:8000
    description: Local development server
paths:
  /users/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      description: Retrieve tasks with advanced filtering and sorting capabilities
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: status
          in: query
          description: Filter by task status (all, pending, completed)
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
        - name: priority
          in: query
          description: Filter by priority level
          schema:
            type: string
            enum: [all, high, medium, low]
            default: all
        - name: tags
          in: query
          description: Filter by tags (comma separated)
          schema:
            type: string
        - name: search
          in: query
          description: Search keyword in title or description
          schema:
            type: string
        - name: sort_by
          in: query
          description: Sort by field
          schema:
            type: string
            enum: [created_at, due_at, priority, title]
            default: due_at
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: limit
          in: query
          description: Number of tasks to return
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          description: Number of tasks to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      tasks:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task'
                      total:
                        type: integer
                      page:
                        type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a new task
      description: Create a new task with advanced features
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /users/{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Retrieve details of a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: Task identifier
          schema:
            type: string
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
    put:
      summary: Update a specific task
      description: Update task details with advanced features
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: Task identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
    delete:
      summary: Delete a specific task
      description: Remove a task from the user's list
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: Task identifier
          schema:
            type: string
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
  /users/{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion status
      description: Toggle completion status of a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: Task identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_completed:
                  type: boolean
                  description: New completion status
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the task
        user_id:
          type: string
          description: Owner of the task
        title:
          type: string
          description: Task title
        description:
          type: string
          description: Task description
        completed:
          type: boolean
          description: Completion status
        priority:
          type: string
          description: Priority level (high, medium, low)
          enum: [high, medium, low]
        tags:
          type: array
          items:
            type: string
          description: Array of tags/categories
        due_at:
          type: string
          format: date-time
          description: Due date and time
        remind_at:
          type: string
          format: date-time
          description: Reminder time
        is_recurring:
          type: boolean
          description: Whether task recurs
        recurrence_pattern:
          type: object
          description: Recurrence pattern details
        next_due_date:
          type: string
          format: date-time
          description: Date of next occurrence for recurring tasks
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Task title
        description:
          type: string
          description: Task description
        priority:
          type: string
          description: Priority level
          enum: [high, medium, low]
          default: medium
        tags:
          type: array
          items:
            type: string
          description: Array of tags
        due_at:
          type: string
          format: date-time
          description: Due date and time
        remind_at:
          type: string
          format: date-time
          description: Reminder time
        is_recurring:
          type: boolean
          description: Whether task should recur
        recurrence_pattern:
          type: object
          description: Recurrence pattern details
        completed:
          type: boolean
          default: false
          description: Initial completion status
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          description: Task title
        description:
          type: string
          description: Task description
        priority:
          type: string
          description: Priority level
          enum: [high, medium, low]
        tags:
          type: array
          items:
            type: string
          description: Array of tags
        due_at:
          type: string
          format: date-time
          description: Due date and time
        remind_at:
          type: string
          format: date-time
          description: Reminder time
        is_recurring:
          type: boolean
          description: Whether task should recur
        recurrence_pattern:
          type: object
          description: Recurrence pattern details
        completed:
          type: boolean
          description: Completion status
    TaskResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Task'